Styling and refining code

+++ delete.py
@@ -12,10 +12,16 @@ class Delete(Handler):
             self.redirect('/blog/login')
 
         else:
-            post_id = self.request.query.split('=')[1]
-            post = Entry.get_by_id(int(post_id))
-            post.delete()
-            # When redirecting, db was still sending me the just deleted item
-            # This workaround avoids that
-            time.sleep(0.1)
-            self.redirect('/blog')
+            user = self.get_user_cookie()
+            author = self.request.get('author')
+            if user.username != author:
+                error = 'Sorry, but you can only delete your own posts!'
+                self.render('error.html', error=error)
+            else:
+                post_id = self.request.query.split('=')[1]
+                post = Entry.get_by_id(int(post_id))
+                post.delete()
+                # When redirecting, db was still sending me the just deleted item
+                # This workaround avoids that
+                time.sleep(0.1)
+                self.redirect('/blog')
